tornado_test/测试/tornado-resources/LehuForum/requirements.txt
requests
bcrypt
redis
tornado
aiomysql
aiofiles
peewee
peewee-async
wtforms
wtforms-tornado
wtforms-json
PyJWT